# openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "HelpMotivateMe.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5001/"
    }
  ],
  "paths": {
...
    "/api/ai/onboarding/chat": {
      "post": {
        "tags": [
          "Ai"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ai/onboarding/transcribe": {
      "post": {
        "tags": [
          "Ai"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/general/chat": {
      "post": {
        "tags": [
          "Ai"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneralChatRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneralChatRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GeneralChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/ai/general/create-identity": {
      "post": {
        "tags": [
          "Ai"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIdentityFromAiRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIdentityFromAiRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIdentityFromAiRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatMessage": {
        "required": [
          "role",
          "content"
        ],
        "type": "object",
        "properties": {
          "role": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        }
      },
      "ChatRequest": {
        "required": [
          "messages",
          "step"
        ],
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            }
          },
          "step": {
            "type": "string"
          },
          "context": {
            "type": [
              "null",
              "object"
            ]
          }
        }
      },
...
      "CreateIdentityFromAiRequest": {
        "required": [
          "name",
          "description",
          "icon",
          "color"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "icon": {
            "type": [
              "null",
              "string"
            ]
          },
          "color": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "GeneralChatRequest": {
        "required": [
          "messages"
        ],
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            }
          },
          "context": {
            "type": [
              "null",
              "object"
            ]
          }
        }
      },
...
      "IdentityResponse": {
        "required": [
          "id",
          "name",
          "description",
          "color",
          "icon",
          "totalTasks",
          "completedTasks",
          "tasksCompletedLast7Days",
          "completionRate",
          "totalGoals",
          "completedGoals",
          "totalProofs",
          "totalDailyCommitments",
          "completedDailyCommitments",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "color": {
            "type": [
              "null",
              "string"
            ]
          },
          "icon": {
            "type": [
              "null",
              "string"
            ]
          },
          "totalTasks": {
            "type": "integer",
            "format": "int32"
          },
          "completedTasks": {
            "type": "integer",
            "format": "int32"
          },
          "tasksCompletedLast7Days": {
            "type": "integer",
            "format": "int32"
          },
          "completionRate": {
            "type": "number",
            "format": "double"
          },
          "totalGoals": {
            "type": "integer",
            "format": "int32"
          },
          "completedGoals": {
            "type": "integer",
            "format": "int32"
          },
          "totalProofs": {
            "type": "integer",
            "format": "int32"
          },
          "totalDailyCommitments": {
            "type": "integer",
            "format": "int32"
          },
          "completedDailyCommitments": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
...
      "TranscriptionResponse": {
        "required": [
          "text",
          "durationSeconds"
        ],
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "durationSeconds": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  },
...
}